Example refactor and README update.
