+++ README.md
@@ -1,20 +1,20 @@
 <img src="http://upload.wikimedia.org/wikipedia/zh/8/82/KKBOX_Logo.png" alt="KKBOX_Logo" align="right" vspace="-50" />
 
-# KKBOX Android Toolkit
+# KKBOX Android Toolkit (a.k.a. KKToolkit)
 
-KKBOX Android Toolkit(KAT) is a set of handy features that our engineer team developed when building KKBOX Android.
+KKBOX Android KKToolkit(KKBOX/android_kktoolkit) is a set of handy features that our engineer team developed when building KKBOX Android.
 It contains wrapped API handling, image manager, custom UIs, notification manager, and others.
 
-## Quick Start 
+## Quick Start
 
 This quick start is intended for app developers. To contribute to the project,
-see [Contributing Code to KAT](https://github.com/KKBOX/KAT/wiki/Contributing-Code-to-KAT).
+see [Contributing Code](https://github.com/KKBOX/android_kktoolkit/wiki/Contributing-Code).
 
 1. Download the latest package, or fork and clone the repository if you like.
 
 1. Include the whole project folder in your Eclipse or ADT.
 
-1. Set kkbox_android_toolkit as a library of your project.
+1. Set android_kktoolkit as a library of your project.
 
 1. Begin coding!
 
@@ -22,9 +22,9 @@ see [Contributing Code to KAT](https://github.com/KKBOX/KAT/wiki/Contributing-Co
 
 You can look at the example app in the tests/ folder for general usages.
 
-If you still have doubt, be sure to check out our [wiki](https://github.com/KKBOX/KAT/wiki/) and see if there is an answer.
+If you still have doubt, be sure to check out our [wiki](https://github.com/KKBOX/android_kktoolkit/wiki/) and see if there is an answer.
 
-There is also an online [doxygen documentation](http://192.168.6.231/doxygen/annotated.html) for you to follow up.
+There is also an online doxygen documentation(coming soon) for you to follow up.
 
 ## Discussion/Forums
 
@@ -32,6 +32,6 @@ Feel free to give any feedback or comments by the issue tracker, or drop us a ma
 
 ## Licensing and Contributions
 
-KKBOX Android Toolkit(KAT) is licensed under a [Apache license](https://github.com/KKBOX/KAT/blob/master/LICENSE.txt). To contribute the project, please see [Contributing](https://github.com/KKBOX/KAT/wiki/Contributing-Code-to-KAT). 
+KKBOX Android KKToolkit(KKToolkit) is licensed under a [Apache license](https://github.com/KKBOX/android_kktoolkit/blob/master/LICENSE.txt). To contribute the project, please see [Contributing](https://github.com/KKBOX/android_kktoolkit/wiki/Contributing-Code).
 
 You may also [join us](http://www.kkbox.com/jobs/) and write anything you need, paid :)
