+++ tests/AndroidManifest.xml
@@ -30,7 +30,7 @@
         android:allowBackup="true"
         android:icon="@drawable/ic_launcher"
         android:label="@string/app_name"
-        android:theme="@style/KKActivityTheme" >
+        android:theme="@style/ExampleActivityTheme" >
         <service android:name="com.kkbox.toolkit.example.ExampleService" />        
         <activity
             android:name="com.kkbox.toolkit.example.MainActivity"
@@ -41,12 +41,12 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name="com.kkbox.toolkit.example.activities.ActivityFragment" />
-        <activity android:name="com.kkbox.toolkit.example.activities.ActivityListFragment" />
-        <activity android:name="com.kkbox.toolkit.example.activities.ActivityTabFragment" />
-        <activity android:name="com.kkbox.toolkit.example.activities.ActivityImage" />
-        <activity android:name="com.kkbox.toolkit.example.activities.ActivityNotification" />
-        <activity android:name="com.kkbox.toolkit.example.activities.ActivityEventQueue" />
+        <activity android:name="com.kkbox.toolkit.example.fragment.ActivityFragment" />
+        <activity android:name="com.kkbox.toolkit.example.listfragment.ActivityListFragment" />
+        <activity android:name="com.kkbox.toolkit.example.tabfragment.ActivityTabFragment" />
+        <activity android:name="com.kkbox.toolkit.example.image.ActivityImage" />
+        <activity android:name="com.kkbox.toolkit.example.notification.ActivityNotification" />
+        <activity android:name="com.kkbox.toolkit.example.eventqueue.ActivityEventQueue" />
     </application>
 
 </manifest>
