+++ tests/res/drawable/selector_button_top_tab_left.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  Copyright (C) 2013 KKBOX Inc.
+
+  Licensed under the Apache License, Version 2.0 (the "License");
+  you may not use this file except in compliance with the License.
+  You may obtain a copy of the License at
+
+  ​http://www.apache.org/licenses/LICENSE-2.0
+
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+-->
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item android:drawable="@drawable/button_top_tab_left_active" android:state_checked="true"/>
+    <item android:drawable="@drawable/button_top_tab_left_pressed" android:state_pressed="true"/>
+    <item android:drawable="@drawable/button_top_tab_left_normal" android:state_checked="false"/>
+
+</selector>
