+++ tests/res/layout-large/listfragment_example.xml
@@ -21,14 +21,14 @@
 
     <fragment
         android:id="@+id/listview_fragment"
-        android:name="com.kkbox.toolkit.example.fragments.ExampleListFragment"
+        android:name="com.kkbox.toolkit.example.listfragment.ExampleListFragment"
         android:layout_width="0dp"
         android:layout_height="match_parent"
         android:layout_weight="1" />
 
     <fragment
         android:id="@+id/message_fragment"
-        android:name="com.kkbox.toolkit.example.fragments.ExampleMessageFragment"
+        android:name="com.kkbox.toolkit.example.listfragment.ExampleMessageFragment"
         android:layout_width="0dp"
         android:layout_height="match_parent"
         android:layout_weight="2" />
