+++ tests/res/layout/eventqueue_example.xml
@@ -25,7 +25,13 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     tools:context=".MainActivity" >
- 
+
+    <Button
+        android:id="@+id/btnAdd"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="@string/addEQ" />
+        
     <Button
         android:id="@+id/btnStart"
         android:layout_width="match_parent"
