+++ tests/res/values/strings.xml
@@ -21,6 +21,7 @@
     <string name="hello_world">Hello world!</string>
 	<string name="tab_1">TabFragment 1</string>
 	<string name="tab_2">TabFragment 2</string>
+	<string name="tab_3">TabFragment 3</string>
 	<string name="change_image">Change Image</string>
 	<string name="imageview">Image View</string>
 	<string name="notify_1">1 Button Dialog</string>
@@ -30,5 +31,6 @@
 	<string name="submit">Submit</string>
 	<string name="response">[Response]</string>
 	<string name="KKBOX">KKBOX</string>
+	<string name="addEQ">Add job</string>
 	<string name="startEQ">Start KKEventQueue</string>
 </resources>
