+++ tests/res/values/styles.xml
@@ -14,5 +14,15 @@
   limitations under the License.
 -->
 <resources>
-
+    <style name="ExampleActivityTheme" parent="KKActivityTheme">
+        <item name="KKTabFragmentStyle">@style/TabFragmentStyle</item>
+    </style>
+    
+    <style name="TabFragmentStyle" parent="KKTabFragmentStyle">
+        <item name="KKTabButtonBackgroundLeft">@drawable/selector_button_top_tab_left</item>
+        <item name="KKTabButtonBackgroundMiddle">@drawable/selector_button_top_tab_center</item>
+        <item name="KKTabButtonBackgroundRight">@drawable/selector_button_top_tab_right</item>
+        <item name="android:height">40dp</item>
+        <item name="android:textSize">14sp</item>
+    </style>
 </resources>
