+++ tests/src/com/kkbox/toolkit/example/listfragment/ExampleListFragment.java
@@ -15,7 +15,7 @@
 /**
  * ExampleListFragment.java: 
  */
-package com.kkbox.toolkit.example.fragments;
+package com.kkbox.toolkit.example.listfragment;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
