+++ tests/src/com/kkbox/toolkit/example/tabfragment/ExampleTabFragment.java
@@ -15,7 +15,7 @@
 /**
  * ExampleTabFragment.java: 
  */
-package com.kkbox.toolkit.example.fragments;
+package com.kkbox.toolkit.example.tabfragment;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -23,13 +23,16 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.kkbox.toolkit.example.R;
+import com.kkbox.toolkit.example.fragment.ExampleFragment;
+import com.kkbox.toolkit.example.listfragment.ExampleListFragment;
+import com.kkbox.toolkit.example.listfragment.ExampleMessageFragment;
 import com.kkbox.toolkit.ui.KKFragment;
 import com.kkbox.toolkit.ui.KKTabFragment;
 import com.kkbox.toolkit.utils.KKDebug;
 
 public class ExampleTabFragment extends KKTabFragment{
 	
-	private int[] buttonTextResourcetId = new int[2];
+	private int[] buttonTextResourcetId = new int[3];
 	private int currentIndex = 0;
 	
 	protected KKFragment onRequestTabFragment(int index, Bundle arguments) {
@@ -43,6 +46,8 @@ public class ExampleTabFragment extends KKTabFragment{
 			case 1:
 				fragment = new ExampleFragment();
 				break;
+			case 2:
+				fragment = new ExampleListFragment();
 		}
 		return fragment;
 	}
@@ -51,6 +56,7 @@ public class ExampleTabFragment extends KKTabFragment{
 		View view = inflater.inflate(R.layout.fragment_tabs, container, false);
 		buttonTextResourcetId[0] = R.string.tab_1;
 		buttonTextResourcetId[1] = R.string.tab_2;
+		buttonTextResourcetId[2] = R.string.tab_3;
 		initView(view, buttonTextResourcetId, false, currentIndex);
 
 		return view;
